Remove KlasaMessage structure
