+++ src/lib/KlasaGlobalMap.js
@@ -1,0 +1,3 @@
+const KlasaGlobalMap = new WeakMap();
+
+module.exports = KlasaGlobalMap;
